+++ retz-client/src/main/java/io/github/retz/cli/CommandGetApp.java
@@ -49,8 +49,7 @@ public class CommandGetApp implements SubCommand {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
 
-        try (Client webClient = new Client(fileConfig.getUri().getHost(),
-                fileConfig.getUri().getPort())) {
+        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
 
             Response res = webClient.getApp(appName);
             LOG.info(res.status());