+++ retz-client/src/main/java/io/github/retz/cli/CommandGetJob.java
@@ -51,8 +51,7 @@ public class CommandGetJob implements SubCommand {
     public int handle(FileConfiguration fileConfig) {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
-        try (Client webClient = new Client(fileConfig.getUri().getHost(),
-                fileConfig.getUri().getPort())) {
+        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
 
             LOG.info("Fetching job detail id={}", id);