+++ retz-client/src/main/java/io/github/retz/cli/CommandKill.java
@@ -47,8 +47,7 @@ public class CommandKill implements SubCommand {
     public int handle(FileConfiguration fileConfig) {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
-        try (Client webClient = new Client(fileConfig.getUri().getHost(),
-                fileConfig.getUri().getPort())) {
+        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
 
             LOG.info("Killing job detail id={}", id);
             Response res = webClient.kill(id);