+++ retz-client/src/main/java/io/github/retz/cli/CommandListApp.java
@@ -44,8 +44,7 @@ public class CommandListApp implements SubCommand {
     public int handle(FileConfiguration fileConfig) {
         LOG.debug("Configuration: {}", fileConfig.toString());
 
-        try (Client webClient = new Client(fileConfig.getUri().getHost(),
-                fileConfig.getUri().getPort())) {
+        try (Client webClient = new Client(fileConfig.getUri(), fileConfig.checkCert())) {
 
             Response res = webClient.listApp();